% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\name{exp}
\alias{exp}
\alias{log}
\title{Exponentiate or logharitmize the value.}
\usage{
exp(x)

log(x)
}
\arguments{
\item{x}{Numeric.}
}
\description{
\code{exp} and \code{log} returns exponentiated or logharitmized values.
}
\details{
\code{exp} and \code{log} functions call the corresponding ones from base
package, but also they set new class to values so that they can be checked
later for they current state. This makes easier to track if the used values
are in their normal or transformated form.
}
\examples{
mini_iris <- iris[c(1, 51, 101), ]
mini_iris
mini_iris$Sepal.Length <- exp(mini_iris$Sepal.Length)
mini_iris$Petal.Width <- log(mini_iris$Petal.Width)
mini_iris
class(mini_iris$Sepal.Length)
class(mini_iris$Petal.Width)
mini_iris$Sepal.Length <- log(mini_iris$Sepal.Length)
mini_iris$Petal.Width <- exp(mini_iris$Petal.Width)
mini_iris
class(mini_iris$Sepal.Length)
class(mini_iris$Petal.Width)

}
