% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lex.R
\name{expect_none}
\alias{expect_none}
\alias{expect_exped}
\alias{expect_loged}
\title{Expect certain object class.}
\usage{
expect_none(x)

expect_exped(x)

expect_loged(x)
}
\arguments{
\item{x}{Object checked for \code{expected} class.}
}
\description{
Returns an error or passed object.
}
\details{
Passed object \code{x} is checked for expected class.
}
\examples{
mini_iris <- iris[c(1, 51, 101), ]
expect_none(mini_iris$Sepal.Length)
mini_iris$log_Sepal.Length <- exp(mini_iris$Sepal.Length)
expect_exped(mini_iris$log_Sepal.Length)
\dontrun{expect_loged(mini_iris$Sepal.Length)}

}
